% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cutoffs}
\alias{cutoffs}
\title{Flood stage cutoffs for USGS stations}
\format{
\subsection{\code{cutoffs}}{

A data frame with 6 rows and 7 columns:
\describe{
\item{stations}{USGS station ID (8-digit)}
\item{major}{Minimum cutoff for major flood stage}
\item{minor}{Minimum cutoff for minor flood stage}
\item{fson}{Whether or not the station has a published flood stage}
\item{minpeaks100}{Cutoff proportional to max peak for only floods to be
reported as peaks}
\item{minpeaks75}{Cutoff proportional to max peak for all floods and peaks
more than 75\% of the flood/minor cutoff to be reported as peaks}
\item{minpeaks50}{Cutoff proportional to max peak for all floods and peaks
more than 50\% of the flood/minor cutoff to be reported as peaks}
}
}
}
\source{
\url{https://www.cnrfc.noaa.gov/}
}
\usage{
cutoffs
}
\description{
Data frame specifying cutoffs for flood stages at six different USGS
stations in the Sierra Nevada.
}
\keyword{datasets}
