% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_detect.R
\name{df_peaks_all}
\alias{df_peaks_all}
\title{df_peaks_all}
\usage{
df_peaks_all(minpeaks_spec = 100)
}
\arguments{
\item{minpeaks_spec}{A number that is equal to 50, 75, or 100. This parameter
specifies the type of flood that the user wants the
peak information for.}
}
\description{
This function makes a dataframe containing peak information for all stations
based off of the flood type parameter
}
\examples{
df_peaks_all(50)
df_peaks_all(75)
}
