% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_detect.R
\name{df_peaks_filt}
\alias{df_peaks_filt}
\title{df_peaks_filt}
\usage{
df_peaks_filt(station, minpeaks_spec = 100)
}
\arguments{
\item{station}{A character or numeric object which specifies which station
the user is interested in. Must be one of the following station ID's:
'10311000', '10310500' , '10309000', '10311200', '10311100', or '10308200'.}

\item{minpeaks_spec}{A numeric object either 100, 75, or 50, which specifies
which peak level to filter for.}
}
\value{
A data frame that includes id, date and time,
flow, and type of flood
}
\description{
This function filters through the dataset peaks_list and filters for the
user specified station, and peak level.
}
