% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_plot.R
\name{peak_plot}
\alias{peak_plot}
\title{Plot streamflow peaks}
\usage{
peak_plot(station, peaks, beg_date, end_date)
}
\arguments{
\item{station}{A character string or numeric object from the following list
of station IDs: "10311000", "10310500" , "10309000", "10311200",
"10311100", or "10308200."}

\item{peaks}{Numeric object from following list: 100, 75, 50. Specifies
amount/type of peaks to include in graph.

\if{html}{\out{<div class="sourceCode">}}\preformatted{100: only include peaks which are floods
75: include all peaks which are floods as well as any peak whose size
is more than 75\% of the flood stage threshold
50: include all peaks which are floods as well as any peak whose size
is more than 50\% of the flood stage threshold
}\if{html}{\out{</div>}}}

\item{beg_date, end_date}{Numeric objects with 10 digits specifying
(in order) the year, month, day, and hour of the beginning and end dates
desired for the plot. For example, March 16, 1997 at 4pm would be designated
by: 1997031616.

Specified start dates must be after the first available date for each
station. Begin dates for periods of records for station data are
listed below:

\if{html}{\out{<div class="sourceCode">}}\preformatted{10311000: 1989-10-01 01:00:00
10310500: 1989-01-19 00:00:00
10309000: 1990-10-01 01:00:00
10311200: 1990-10-01 01:00:00
10311100: 1990-10-01 01:00:00
10308200: 1990-10-01 01:00:00
}\if{html}{\out{</div>}}}
}
\description{
Function to plot peak streamflows for user-specified USGS stations and time
periods.
}
