% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_generics.R
\name{plot.peakflows}
\alias{plot.peakflows}
\title{Plot the Function}
\usage{
\method{plot}{peakflows}(station, peaks, beg_date, end_date, x, y, ...)
}
\arguments{
\item{station}{A character or numeric object which specifies which station
the user is interested in. Must be one of the following station ID's:
'10311000', '10310500' , '10309000', '10311200', '10311100', or '10308200'.}

\item{peaks}{Not sure yet}

\item{beg_date}{An object of class Date-Time representing the beginning date
of when to include the data in the graph represented in the format
YYYY-MM-DD.}

\item{end_date}{An object of class Date-Time representing the end date of
when to include the data in the graph represented in the format
YYYY-MM-DD.}

\item{x}{An object required to run the function}

\item{y}{An object required to run the function}

\item{...}{Other parameters}
}
\description{
Plot the Function
}
