% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rhv_tot}
\alias{rhv_tot}
\title{USGS hourly streamflow data}
\format{
\subsection{\code{rhv_tot}}{

A data frame with 1,734,174 rows and 3 columns:
\describe{
\item{id}{USGS station ID (8-digit)}
\item{datetime}{Hourly reported date and time of streamflow measurement}
\item{max_flow}{Streamflow measurement (cfs)}
}
}
}
\source{
\url{https://waterdata.usgs.gov/nwis/}
}
\usage{
rhv_tot
}
\description{
A subset of data from the USGS website for six different stream gauges in
the Sierra Nevada.
}
\keyword{datasets}
